industrial-metaverse-amr/industrial_metaverse/visualization/receive/requirements.txt
eclipse-zenoh==0.10.0rc0
pyyaml==6.0.1
ffmpeg-python==0.2.0
fire==0.5.0
